rename android-feature-views module to android-feature-screens
