+++ android/features/android-feature-screens/proguard-rules.pro
