+++ android/features/android-feature-screens/src/main/AndroidManifest.xml
