+++ android/features/android-feature-screens/src/main/java/io/matthewnelson/android_feature_screens/ui/motionlayout/MotionLayoutViewState.kt
@@ -13,7 +13,7 @@
 *  See the License for the specific language governing permissions and
 *  limitations under the License.
 * */
-package io.matthewnelson.android_feature_views.ui.motionlayout
+package io.matthewnelson.android_feature_screens.ui.motionlayout
 
 import androidx.annotation.IdRes
 import androidx.constraintlayout.motion.widget.MotionLayout
