+++ android/features/android-feature-screens/src/main/java/io/matthewnelson/android_feature_screens/ui/sideeffect/SideEffectViewModel.kt
@@ -13,12 +13,12 @@
 *  See the License for the specific language governing permissions and
 *  limitations under the License.
 * */
-package io.matthewnelson.android_feature_views.ui.sideeffect
+package io.matthewnelson.android_feature_screens.ui.sideeffect
 
 import io.matthewnelson.concept_views.sideeffect.SideEffect
 import io.matthewnelson.concept_views.sideeffect.SideEffectContainer
 import io.matthewnelson.concept_views.sideeffect.collect
-import io.matthewnelson.android_feature_views.ui.base.BaseViewModel
+import io.matthewnelson.android_feature_screens.ui.base.BaseViewModel
 import io.matthewnelson.concept_views.viewstate.ViewState
 
 suspend inline fun <T, SE: SideEffect<T>, VS: ViewState<VS>> SideEffectViewModel<T, SE, VS>.collectSideEffects(
