+++ android/features/android-feature-screens/src/main/java/io/matthewnelson/android_feature_screens/util/ViewExtensions.kt
@@ -13,7 +13,7 @@
 *  See the License for the specific language governing permissions and
 *  limitations under the License.
 * */
-package io.matthewnelson.android_feature_views.util
+package io.matthewnelson.android_feature_screens.util
 
 import android.content.res.Resources
 import android.view.View
