+++ android/features/android-feature-screens/src/main/res/values/strings.xml
